requirements.txt
asgiref==3.7.2
cfgv==3.3.1
distlib==0.3.6
Django==4.2.2
django-guardian==2.4.0
dnspython==1.16.0
filelock==3.12.2
identify==2.5.24
nodeenv==1.8.0
platformdirs==3.8.0
psycopg2-binary==2.9.6
pymongo==3.11.0
python-dotenv==1.0.0
PyYAML==6.0
sqlparse==0.4.4
virtualenv==20.23.1
